% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind.R
\name{ind}
\alias{ind}
\title{Individualize function}
\usage{
ind(data)
}
\arguments{
\item{data}{dataset you want to add "individual" column to}
}
\value{
The original output plus a row number variable
}
\description{
Individualize function
}
\examples{
ind(iris)
}
