% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_al_k.R
\name{missing_al_k}
\alias{missing_al_k}
\title{We can see all observations who are missing at least k variables}
\usage{
missing_al_k(.data, k = 0)
}
\arguments{
\item{.data}{name of dataset}

\item{k}{number of missing values}
}
\value{
list of individuals who are missing at least k values
}
\description{
We can see all observations who are missing at least k variables
}
\examples{
missing_al_k(iris, k=2)
}
