% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_k.R
\name{missing_k}
\alias{missing_k}
\title{We can see all observations who are missing exactly k variables}
\usage{
missing_k(.data, k = 0)
}
\arguments{
\item{.data}{dataset of interest}

\item{k}{number of missing values}
}
\value{
list of individuals who are missing k values
}
\description{
We can see all observations who are missing exactly k variables
}
\examples{
missing_k(iris, k=2)
}
