% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missoutput.R
\name{missoutput}
\alias{missoutput}
\title{This puts together individual, # missing, and list of missing variables}
\usage{
missoutput(df, n = 1)
}
\arguments{
\item{df}{data frame of interest--should be individualized}

\item{n}{number of missing values of interest}
}
\value{
individual, number of missing values, variables missing
}
\description{
This puts together individual, # missing, and list of missing variables
}
\examples{
air <- ind(airquality)
missoutput(air, n=1)
}
