% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{prompt}
\alias{prompt}
\title{here is a function that writes an R markdown for you
you will enter the parameters in the datasets.R file
it will write a report.txt file to your directory
the text can be copied into an RMD for an instant report}
\usage{
prompt(data, people = "all", variables = "all")
}
\arguments{
\item{data}{dataset of interest--should be initialized}

\item{people}{vector of individual row numbers}

\item{variables}{vector of columns of interest in form c(1,2,...)}
}
\value{
dataset with specified people and columns, RMD template text file
}
\description{
here is a function that writes an R markdown for you
you will enter the parameters in the datasets.R file
it will write a report.txt file to your directory
the text can be copied into an RMD for an instant report
}
\examples{
prompt(iris, c(2,4,5,6,7,8), c(1,2,3))
}
